backend/predictor/pyproject.toml
[tool.poetry]
name = "predictor"
version = "0.1.0"
description = "FastAPI predictor service"
authors = ["Your Name <you@example.com>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "3.11.13"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
torch = "2.1.1"
torchvision = "0.16.1"
pillow = "^10.2.0"
asyncpg = "^0.30.0"
python-dotenv = "^1.1.1"
aiofiles = "^24.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
